extract new component to individual file
